Add signal information in scan results for vanilla Ralink drivers.
Vanilla Ralink drivers don't report signal information. In this case, we get
scan results from the WEXT interface if available, and merge signal information
into the original scan results. That should work with all Ralink drivers as
long as they have support for the WEXT interface. That removes the need to
change Ralink drivers. Besides, it's not even possible (or would be difficult)
with some driver versions as scan results are directly generated by the
firmware and passed as-is to the user space.